requirements.txt
langchain
langchain_community
langchain_core
langchain_openai
python-dotenv
streamlit
wordcloud
matplotlib
scikit-learn
pandas
crewai
